+++ src/colvarproxy.cpp
@@ -63,7 +63,7 @@ bool colvarproxy_system::total_forces_same_step() const
 
 inline int round_to_integer(cvm::real x)
 {
-  return cvm::floor(x+0.5);
+  return int(cvm::floor(x+0.5));
 }
